move App Protect WAF controller code to it's own file
